fix: change vuejs3 devtools to vuejs devtools
vuejs3 devtools points to the beta track of the vuejs devtools for
vuejs devtools version 6. Version 6 with support for vue.js 3 is now out
of beta and installable via vuejs devtools. This fixes a bug in dev mode
where the vue-router does not work if the vue devtools have been opened
at any time