fix(vuex user state): fix user duplication bug
fix a bug where users are erroneously inserted into the users array
instead the existing user being updated if the username is changed.
This is caused by the addOrUpdateUser method checking if the usernames
are the same for determining object equality. The fix is to just use
the id for checking object equality