feature(vuex-states): add pagination support
Add pagination support the the vuex-state modules. This concists of a
page and total variable in the state. The page variable always hold the
last retrieved page by the paginated call. We assume that there will
be only one paginated view per state at one time. The total variable
indicates the total amount of retrievable entities. This is used for
calculating when the last available entities have been loaded. It is
also updated with every paginated call in a state.