refactor: change table components and group views
Change prop names in the table components from entity to data. Change
the mapping of operation ids to operation titles to occur only when
displayed in AllGroups view. Change style of CreateNewGroup and
EditGroup view and add use of FloatingModal component with new selection
of members based on table components, a search bar and tags to delete
members even if the table is on another page