fix(login): fix logged in user not retrieved
Fixes a bug in the login code where the user which is logging in is
not loaded on login. This leads to problems when checking permissions
as the is_admin property must be readable. Was masked until now by the
loading of users, operations and groups in the mainapplication component